+++ src/main/java/de/maxhenkel/car/integration/waila/HUDHandlerGenerator.java
@@ -5,16 +5,13 @@ import de.maxhenkel.car.blocks.tileentity.TileEntityGenerator;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import snownee.jade.api.BlockAccessor;
 import snownee.jade.api.IBlockComponentProvider;
 import snownee.jade.api.IServerDataProvider;
 import snownee.jade.api.ITooltip;
 import snownee.jade.api.config.IPluginConfig;
 
-public class HUDHandlerGenerator implements IBlockComponentProvider, IServerDataProvider<BlockEntity> {
+public class HUDHandlerGenerator implements IBlockComponentProvider, IServerDataProvider<BlockAccessor> {
 
     public static final HUDHandlerGenerator INSTANCE = new HUDHandlerGenerator();
 
@@ -32,8 +29,8 @@ public class HUDHandlerGenerator implements IBlockComponentProvider, IServerData
     }
 
     @Override
-    public void appendServerData(CompoundTag compoundTag, ServerPlayer serverPlayer, Level level, BlockEntity blockEntity, boolean b) {
-        TileEntityGenerator generator = (TileEntityGenerator) blockEntity;
+    public void appendServerData(CompoundTag compoundTag, BlockAccessor blockAccessor) {
+        TileEntityGenerator generator = (TileEntityGenerator) blockAccessor.getBlockEntity();
         compoundTag.putInt("energy", generator.getEnergyStored());
         compoundTag.putInt("max_energy", generator.getMaxEnergyStored());
         compoundTag.putInt("fluid", generator.getFluidInTank(0).getAmount());